Initial commit
This project was adapted to work with Unity 6 and Universal Render Pipeline.
It contains the basic game, a maze where the objects and player will be instantiated at random, first person controller from the "GDT Solutions for Unity" repo and a basic system that allows you to change the language.